+++ .storybook/preview.cjs
@@ -1,4 +1,4 @@
-import '../src/assets/index.css'
+import '../src/index.css'
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
