--- src/stories/Header.stories.js
@@ -1,34 +1,0 @@
-import MyHeader from "./Header.vue";
-
-export default {
-  title: "Example/Header",
-  component: MyHeader,
-  parameters: {
-    // More on Story layout: https://storybook.js.org/docs/vue/configure/story-layout
-    layout: "fullscreen",
-  },
-};
-
-const Template = (args) => ({
-  // Components used in your story `template` are defined in the `components` object
-  components: { MyHeader },
-  // The story's `args` need to be mapped into the template through the `setup()` method
-  setup() {
-    // Story args can be spread into the returned object
-    return { ...args };
-  },
-  // Then, the spread values can be accessed directly in the template
-  template: '<my-header :user="user" />',
-});
-
-export const LoggedIn = Template.bind({});
-LoggedIn.args = {
-  user: {
-    name: "Jane Doe",
-  },
-};
-
-export const LoggedOut = Template.bind({});
-LoggedOut.args = {
-  user: null,
-};
