--- src/stories/index.js
@@ -1,5 +1,0 @@
-import "tailwindcss/tailwind.css";
-
-import Button from "./Button2.vue";
-
-export { Button };
